chore :: [#196] SectionView Component로 이동
